Renamed Dashboard to WeeklyPlanner
The auto generated code had dashboard instead of WeeklyPlanner this commit updates that
