+++ app/src/main/java/com/example/prikkie/ui/weekly_planner/WeeklyPlannerFragment.java
@@ -1,4 +1,4 @@
-package com.example.prikkie.ui.dashboard;
+package com.example.prikkie.ui.weekly_planner;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -14,17 +14,17 @@ import androidx.lifecycle.ViewModelProviders;
 
 import com.example.prikkie.R;
 
-public class DashboardFragment extends Fragment {
+public class WeeklyPlannerFragment extends Fragment {
 
-    private DashboardViewModel dashboardViewModel;
+    private WeeklyPlannerViewModel weeklyPlannerViewModel;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        dashboardViewModel =
-                ViewModelProviders.of(this).get(DashboardViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_dashboard, container, false);
-        final TextView textView = root.findViewById(R.id.text_dashboard);
-        dashboardViewModel.getText().observe(this, new Observer<String>() {
+        weeklyPlannerViewModel =
+                ViewModelProviders.of(this).get(WeeklyPlannerViewModel.class);
+        View root = inflater.inflate(R.layout.fragment_weekly_planner, container, false);
+        final TextView textView = root.findViewById(R.id.text_weekly_planner);
+        weeklyPlannerViewModel.getText().observe(this, new Observer<String>() {
             @Override
             public void onChanged(@Nullable String s) {
                 textView.setText(s);
