+++ app/src/main/java/com/example/prikkie/ui/weekly_planner/WeeklyPlannerViewModel.java
@@ -1,16 +1,16 @@
-package com.example.prikkie.ui.dashboard;
+package com.example.prikkie.ui.weekly_planner;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class WeeklyPlannerViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public WeeklyPlannerViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is the weekly planner fragment");
     }
 
     public LiveData<String> getText() {
