+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,9 +7,9 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_dashboard"
+        android:id="@+id/navigation_weekly_planner"
         android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:title="@string/title_weekly_planner" />
 
     <item
         android:id="@+id/navigation_notifications"
