+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -12,10 +12,10 @@
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.example.prikkie.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        android:id="@+id/navigation_weekly_planner"
+        android:name="com.example.prikkie.ui.weekly_planner.WeeklyPlannerFragment"
+        android:label="@string/title_weekly_planner"
+        tools:layout="@layout/fragment_weekly_planner" />
 
     <fragment
         android:id="@+id/navigation_notifications"
