+++ app/src/main/res/values/strings.xml
@@ -9,7 +9,7 @@
 
     <string name="title_activity_main_nav">MainActivityNav</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Weekly Planner</string>
+    <string name="title_weekly_planner">Weekly Planner</string>
     <string name="title_notifications">Multiplayer</string>
     <string name="title_account">Account</string>
 </resources>
